# gpt-config/openapi-actions.yaml
openapi: "3.1.0"
info:
  title: Course Companion FTE – Actions API
  description: >
    FastAPI backend for the Course Companion FTE ChatGPT Custom GPT.
    Provides course content, quiz submission, progress tracking, and search.
    All endpoints require a Bearer JWT token (obtained from /auth/login).
    Zero LLM calls happen on the backend — all intelligence is in this GPT.
  version: "1.0.0"
servers:
  - url: https://your-backend-domain.com/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "student@example.com"
        password:
          type: string
          minLength: 8
          example: "securepassword123"
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (1 hour expiry)
        refresh_token:
          type: string
          description: JWT refresh token (7 day expiry)
        token_type:
          type: string
          example: "bearer"
        user:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
        tier:
          type: string
          enum: [free, pro]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    ChapterSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: integer
          description: Chapter number (1-5)
        slug:
          type: string
          example: "what-is-generative-ai"
        title:
          type: string
        summary:
          type: string
          description: One-sentence chapter description
        estimated_mins:
          type: integer
          description: Estimated reading time in minutes
        tier_required:
          type: string
          enum: [free, pro]
        has_quiz:
          type: boolean
        tags:
          type: array
          items:
            type: string
    ChapterDetail:
      allOf:
        - $ref: '#/components/schemas/ChapterSummary'
        - type: object
          properties:
            content_md:
              type: string
              description: Full chapter content in Markdown format
...
    SearchResult:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            type: object
            properties:
              chapter_id:
                type: string
              chapter_number:
                type: integer
              title:
                type: string
              excerpt:
                type: string
                description: Relevant excerpt with search terms highlighted
              relevance_score:
                type: number
        total:
          type: integer
        searches_today:
          type: integer
        searches_remaining:
          type: integer
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      summary: Create a new user account
      description: Register with email and password. Returns JWT token pair immediately — no email verification required.
      security: []
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        "201":
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "400":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /chapters/{chapter_id}:
    get:
      operationId: getChapter
      summary: Get full chapter content
      description: >
        Returns the complete chapter including full markdown content.
        Free users can only access chapters 1-3. Returns 403 for locked chapters.
        Call this when the user wants to read a specific chapter.
      tags: [Chapters]
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
          description: Chapter UUID or slug (e.g. "what-is-generative-ai")
      responses:
        "200":
          description: Full chapter detail with content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterDetail'
        "403":
          description: Pro tier required
        "404":
          description: Chapter not found
...
  /search:
    get:
      operationId: searchChapters
      summary: Full-text search across all chapters
      description: >
        Search chapter content using PostgreSQL full-text search (pg_tsvector).
        Returns ranked results with relevant excerpts.
        Free users limited to 10 searches per day.
        Use this when the user asks to find information or says "search for...".
      tags: [Search]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Search query (e.g. "attention mechanism", "RAG", "fine-tuning")
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 5
            maximum: 20
      responses:
        "200":
          description: Search results with excerpts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
        "429":
          description: Daily search limit reached (free tier)

  # ── Health ─────────────────────────────────────────────────────────────────
...